refactor: delete unused component auth_input
