--- src/components/login_register_inputs/auth_input.tsx
@@ -1,57 +1,0 @@
-import Visibility from '@mui/icons-material/Visibility';
-import VisibilityOff from '@mui/icons-material/VisibilityOff';
-import { IconButton, InputAdornment, TextField } from '@mui/material';
-import { useState } from 'react';
-import { Controller, useFormContext } from 'react-hook-form';
-
-interface AuthInputProps {
-  name: 'email' | 'password';
-  label: string;
-}
-
-const AuthInput = ({ name, label }: AuthInputProps) => {
-  const {
-    control,
-    formState: { errors },
-  } = useFormContext();
-  const [showPassword, setShowPassword] = useState(false);
-
-  const handleTogglePassword = () => {
-    setShowPassword(prev => !prev);
-  };
-
-  return (
-    <Controller
-      control={control}
-      name={name}
-      render={({ field }) => (
-        <TextField
-          {...field}
-          error={!!errors[name]}
-          helperText={errors[name]?.message?.toString()}
-          label={label}
-          size="small"
-          slotProps={
-            name === 'password'
-              ? {
-                  input: {
-                    endAdornment: (
-                      <InputAdornment position="end">
-                        <IconButton edge="end" onClick={handleTogglePassword}>
-                          {showPassword ? <Visibility /> : <VisibilityOff />}
-                        </IconButton>
-                      </InputAdornment>
-                    ),
-                  },
-                }
-              : undefined
-          }
-          // sx={{ width: '100%' }}
-          type={name === 'password' && !showPassword ? 'password' : 'text'}
-          variant="outlined"
-        />
-      )}
-    />
-  );
-};
-export default AuthInput;
