+++ src/pages/shop/components/filter/utils/extract-price-range.tsx
